[VPS-033] Add useGraph hook
Moved graph model to the frontend to make it easier for other members of the team to use. Defined a new pi/scenario/:scenarioId/scene/all endpoint that returns an array of all scenes with their full scene data. Also created a useGraph hook to add to the usability of the model.