SpeechTextComponents can be deleted from canvas
div element in the new SpeechTextComponent was the first element to trigger the keypress event because it was focussable. This meant that the event.target.tagName was "DIV" & the deleteElement() function isn't called 🥲

to solve this issue 🛠️🛠️:
capture keydown event (delete) on body so event is triggered (& captured) by the body element.