fix: validate group data csv upload (110) (#278)
* Make group creation error handling work

Current error handling doesn't actually do anything since the responses aren't returned (in fact, it crashes the server).

Fixed this issue by adding return statements. Also had to change the foreach loops to normal for loops so that the returns would work as expected.

* Updated return information for creating groups

Issues with inputs should return bad request and not no content. Also updated error messages to hopefully be a little more informative for end user.

* Updated error messaging in manage groups page

Should hopefully be more information for users now

* Extended toast timer on manage groups page

3s felt too short to read the whole error message. 5s should hopefully be more comfortable for users.

* Updated HTTP status for create groups

Probably makes more sense to be bad request than to be conflict. We seem to validate only on the data that's passed and not on the data being stored.

Also made a quick update to one of the messages (wasn't happy with what I wrote before)

* Fixed formatting issues

I forgot to install prettier 😔